requirements.txt
django.txt
[pip install django]

psycopg2
[pip install psycopg2]

psycopg2-binary
[pip install psycopg2-binary]

django-autoslug
[pip install django-autoslug]

pillow
[pip install pillow]

*may need to download visual studio build tools for psycopg*
[https://visualstudio.microsoft.com/thank-you-downloading-visual-studio/?sku=BuildTools&rel=16]